pyproject.toml
[project]
name = "mcp-bq-sse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigquery>=3.27.0",
    "mcp>=1.0.0",
    "uvicorn>=0.24.0",
    "starlette>=0.27.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-bq-sse = "mcp_bq_sse:main"

[tool.uv]
dev-dependencies = [
    "black>=23.0.0",
    "isort>=5.0.0",
    "ruff>=0.1.0",
]